pipeline: rename `hash_session` to `hash_driver` and re-organise interface
Renamed `hash_pipeline::Compiler` to `hash_driver::driver::Driver` and
implemented the `CompilerInterface` for this type. This will now be used
as the API of the compiler, enabling external users to use the compiler
through the `hash_driver` crate.
