+++ compiler/hash-driver/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hash-session"
+name = "hash-driver"
 version = "0.1.0"
 authors = ["The Hash Language authors"]
 edition = "2021"
