+++ compiler/hash-driver/src/lib.rs
@@ -7,9 +7,11 @@
 //! methods for "selecting" the information that is needed by the stage.
 //! This creates a clear separation between the stages and the global state,
 //! keeping the crate dependency graph clean.
+pub mod driver;
 
 use std::collections::HashSet;
 
+use driver::Driver;
 use hash_ast::node_map::NodeMap;
 use hash_ast_desugaring::{AstDesugaringCtx, AstDesugaringCtxQuery, AstDesugaringPass};
 use hash_ast_expand::{AstExpansionCtx, AstExpansionCtxQuery, AstExpansionPass};
@@ -25,7 +27,6 @@ use hash_pipeline::{
     interface::{CompilerInterface, CompilerOutputStream, CompilerStage},
     settings::CompilerSettings,
     workspace::Workspace,
-    Compiler,
 };
 use hash_reporting::report::Report;
 use hash_semantics::{
@@ -43,17 +44,15 @@ pub struct CompilerBuilder;
 
 impl CompilerBuilder {
     /// Create a new [Compiler] with the default stage configuration.
-    pub fn build_with_settings(settings: CompilerSettings) -> Compiler<DefaultCompilerInterface> {
-        let session = utils::emit_on_fatal_error(|| DefaultCompilerInterface::new(settings));
+    pub fn build_with_settings(settings: CompilerSettings) -> Driver<Compiler> {
+        let session = utils::emit_on_fatal_error(|| Compiler::new(settings));
         Self::build_with_interface(session)
     }
 
     /// Create a new [Compiler] with a specified compiler interface that
     /// supports the [DefaultCtxQuery] trait, i.e. all of the stages within
     /// the standard pipeline.
-    pub fn build_with_interface<I: CompilerInterface + DefaultCtxQuery>(
-        interface: I,
-    ) -> Compiler<I> {
+    pub fn build_with_interface<I: CompilerInterface + DefaultCtxQuery>(interface: I) -> Driver<I> {
         Self::build(
             interface,
             vec![
@@ -77,8 +76,8 @@ impl CompilerBuilder {
     pub fn build<I: CompilerInterface>(
         ctx: I,
         stages: Vec<Box<dyn CompilerStage<I>>>,
-    ) -> Compiler<I> {
-        let mut compiler = Compiler::new(ctx, stages);
+    ) -> Driver<I> {
+        let mut compiler = Driver::new(ctx, stages);
 
         compiler.bootstrap();
         compiler
@@ -122,7 +121,7 @@ pub mod utils {
 /// compiler instance. Each stage of the compiler contains a `State` type
 /// parameter which the compiler stores so that incremental executions of the
 /// compiler are possible.
-pub struct DefaultCompilerInterface {
+pub struct Compiler {
     /// The collected workspace sources for the current job.
     pub workspace: Workspace,
 
@@ -170,7 +169,7 @@ pub struct DefaultCompilerInterface {
     pub layout_storage: LayoutCtx,
 }
 
-impl DefaultCompilerInterface {
+impl Compiler {
     /// Create a new compiler session with the input and output streams being
     /// set as the standard error and standard output streams.
     pub fn new(settings: CompilerSettings) -> Result<Self, PipelineError> {
@@ -225,9 +224,9 @@ impl DefaultCompilerInterface {
     }
 }
 
-impl DefaultCtxQuery for DefaultCompilerInterface {}
+impl DefaultCtxQuery for Compiler {}
 
-impl CompilerInterface for DefaultCompilerInterface {
+impl CompilerInterface for Compiler {
     /// Get a reference to the error [CompilerOutputStream].
     fn error_stream(&self) -> CompilerOutputStream {
         (self.error_stream)()
@@ -235,6 +234,13 @@ impl CompilerInterface for DefaultCompilerInterface {
 
     /// Get a reference to the output [CompilerOutputStream].
     fn output_stream(&self) -> CompilerOutputStream {
+        // @@Craziness: rust syntax failure (RSF) is causing the compiler to
+        // think that the `output_stream` field is a function, so we have to
+        // wrap it in parentheses to get around this.
+        //
+        // See:
+        //
+        // (This was generated by GitHub Copilot)
         (self.output_stream)()
     }
 
@@ -279,25 +285,25 @@ impl CompilerInterface for DefaultCompilerInterface {
     }
 }
 
-impl ParserCtxQuery for DefaultCompilerInterface {
+impl ParserCtxQuery for Compiler {
     fn data(&mut self) -> ParserCtx {
         ParserCtx { workspace: &mut self.workspace, pool: &self.pool }
     }
 }
 
-impl AstDesugaringCtxQuery for DefaultCompilerInterface {
+impl AstDesugaringCtxQuery for Compiler {
     fn data(&mut self) -> AstDesugaringCtx {
         AstDesugaringCtx { workspace: &mut self.workspace, pool: &self.pool }
     }
 }
 
-impl UntypedSemanticAnalysisCtxQuery for DefaultCompilerInterface {
+impl UntypedSemanticAnalysisCtxQuery for Compiler {
     fn data(&mut self) -> UntypedSemanticAnalysisCtx {
         UntypedSemanticAnalysisCtx { workspace: &mut self.workspace, pool: &self.pool }
     }
 }
 
-impl AstExpansionCtxQuery for DefaultCompilerInterface {
+impl AstExpansionCtxQuery for Compiler {
     fn data(&mut self) -> AstExpansionCtx {
         let output_stream = self.output_stream();
         AstExpansionCtx {
@@ -308,7 +314,7 @@ impl AstExpansionCtxQuery for DefaultCompilerInterface {
     }
 }
 
-impl SemanticAnalysisCtxQuery for DefaultCompilerInterface {
+impl SemanticAnalysisCtxQuery for Compiler {
     fn data(&mut self) -> SemanticAnalysisCtx {
         SemanticAnalysisCtx {
             workspace: &mut self.workspace,
@@ -324,7 +330,7 @@ impl SemanticAnalysisCtxQuery for DefaultCompilerInterface {
     }
 }
 
-impl LoweringCtxQuery for DefaultCompilerInterface {
+impl LoweringCtxQuery for Compiler {
     fn data(&mut self) -> LoweringCtx {
         let output_stream = self.output_stream();
         LoweringCtx {
@@ -339,7 +345,7 @@ impl LoweringCtxQuery for DefaultCompilerInterface {
     }
 }
 
-impl BackendCtxQuery for DefaultCompilerInterface {
+impl BackendCtxQuery for Compiler {
     fn data(&mut self) -> BackendCtx {
         let output_stream = self.output_stream();
 
@@ -355,7 +361,7 @@ impl BackendCtxQuery for DefaultCompilerInterface {
     }
 }
 
-impl LinkerCtxQuery for DefaultCompilerInterface {
+impl LinkerCtxQuery for Compiler {
     fn data(&mut self) -> hash_link::LinkerCtx<'_> {
         let stdout = self.output_stream();
