+++ compiler/hash/Cargo.toml
@@ -19,7 +19,7 @@ hash-alloc = { path = "../hash-alloc" }
 hash-ast = { path = "../hash-ast" }
 hash-pipeline = { path = "../hash-pipeline" }
 hash-reporting = { path = "../hash-reporting" }
-hash-session = { path = "../hash-session" }
+hash-driver = { path = "../hash-driver" }
 hash-source = { path = "../hash-source" }
 hash-target = { path = "../hash-target" }
 hash-utils = { path = "../hash-utils" }
